+++ libs/ui/src/lib/callout/index.ts
@@ -1,5 +1,6 @@
 export { NbCallout } from './nb-callout';
 export type {
+  NbCalloutBorder,
   NbCalloutLayout,
   NbCalloutRadius,
   NbCalloutShadow,