+++ libs/ui/src/lib/cluster/index.ts
@@ -4,6 +4,7 @@ export type {
   NbClusterAlign,
   NbClusterGap,
   NbClusterJustify,
+  NbClusterPadding,
   NbClusterSeparator,
   NbClusterWrap,
 } from './nb-cluster';