+++ libs/ui/src/lib/input-group/index.ts
@@ -1,4 +1,5 @@
 export { NbInputGroup } from './nb-input-group';
+export type { NbInputAffixAlign } from './input-group.types';
 export { NbInputPrefix } from './nb-input-group-prefix';
 export type { NbInputPrefixAlign } from './nb-input-group-prefix';
 export { NbInputSuffix } from './nb-input-group-suffix';